+++ src/Stolons/Models/Product.cs
@@ -304,7 +304,7 @@ namespace Stolons.Models
             Unlimited = 2
         }
 
-        public string GetQuantityString(int quantity)
+        public string GetQuantityString(decimal quantity)
         {
             if (Type == SellType.Piece)
             {
@@ -319,7 +319,7 @@ namespace Stolons.Models
             }
             else
             {
-                float qty = (quantity * QuantityStep);
+                decimal qty = (quantity * QuantityStep);
                 string strUnit;
                 if (ProductUnit == Product.Unit.Kg)
                 {
